fix(web): mirror NEP-50 float32 narrowing in radiusAt; test what fixtures cannot
place.py:108 computes `r_max - (r_max - r_min) * risk` in float32: under NEP 50
a Python float is a weak scalar, so numpy narrows it to the array's dtype
rather than promoting the array. JavaScript computed the same expression in
float64. Measured over the whole detectKm slider on the committed Los Padres
rasters: agreement at the shipped 2.0 km default, but 14799 vs 14794 candidates
at 0.5 km and 10262 vs 10322 at 0.6 km -- different networks, reachable by
dragging one slider.

radiusAt now mirrors the narrowing with Math.fround. The span must be narrowed
BEFORE the multiply; verified against numpy 2.4.4 that keeping it at float64
disagrees wherever r_max - r_min is not float32-representable (detectKm 0.7,
1.3, 3.3, 4.9). After the fix the sweep is 0/46 diverging and the worst
coordinate deviation is 2.8e-14 km.

export_fixtures.py now hands place.py genuine float32 arrays rather than
float32 values in a float64 array, so the golden fixtures pin the arithmetic
the browser actually runs. poisson.json and cover.json regenerated;
refrng.json and frame.json are byte-identical.

Also close the numerical-machinery coverage gap. cover.json's weights are
float32, hence multiples of 2^-26, so its sums are exact and order-independent
-- dropping the CELF 1e-12, deleting the ascending hit-list sort, or replacing
pairwiseSum with a naive loop all shipped green with zero delta. Test the
predicates directly instead (celfAcceptsRefresh, seenDemand, pairwiseSum vs
naive), add n=130 to sum.test.ts where the 8-alignment guard first bites, pin
gridShape's actual shapes, and add layers.test.ts for the vertical flip and the
half-extent coordinate correction.

+++ tools/bake/export_fixtures.py
@@ -31,7 +31,7 @@ def _repo_geo_and_place():
 # maximizes the per-draw rejection probability at this magnitude and keeps
 # the expected number of draws-to-first-rejection near 2**19 (~524k) instead
 # of the 2**24+ (~16.7M) a naively chosen HIGH near 2**45 would need.
-REJECTION_HIGH = 35184372088833  # 2**45 + 524289, tuned for max threshold/HIGH
+REJECTION_HIGH = 35184372088833  # 2**45 + 1, tuned for max threshold/HIGH
 REJECTION_SEED = 999_999
 REJECTION_RECORD = 8
 REJECTION_SEARCH_BOUND = 5_000_000
@@ -246,14 +246,20 @@ def poisson_fixture() -> None:
             # instead of relying on a second sort implementation to agree
             # with numpy's unstable tie-break.
 
-        # web/src/lib/types.ts stores Field.data as Float32Array. Snap here
-        # so place.py computes on the exact values the TS port will read
-        # back -- otherwise every radius and candidate position drifts by
-        # the float64-vs-float32 rounding gap, and that drift compounds
-        # across the active-list chain as points get placed. Confirmed as
-        # the standing rule for fixtures that feed a Field: the reference
-        # must compute on the exact values the consumer will hold.
-        risk = risk.astype(np.float32).astype(np.float64)
+        # web/src/lib/types.ts stores Field.data as Float32Array, and
+        # risk.bin ships as "<f4", so keep this array float32 -- dtype and
+        # all, not merely float32-*valued*.
+        #
+        # The dtype matters as much as the values. Under NEP 50 (numpy >=
+        # 2.0) a Python float is a weak scalar, so place.py:108's
+        # `r_max_km - (r_max_km - r_min_km) * _sample_field(risk, ...)`
+        # narrows to float32 when `risk` is float32 and stays float64 when it
+        # is not. A float64 array carrying float32 values would therefore
+        # make this fixture pin the *wrong* arithmetic: the reference would
+        # compute radii in double while the browser computes them in single.
+        # web/src/lib/field.ts mirrors the narrowing with Math.fround; this
+        # is the reference that holds it honest.
+        risk = risk.astype(np.float32)
 
         mask = None
         if mask_kind == "exclude_top_bottom_rows":
@@ -301,12 +307,14 @@ def cover_fixture() -> None:
     v = yy / (ny - 1)
     risk = np.clip(0.5 + 0.5 * np.sin(3 * np.pi * u) * np.cos(2 * np.pi * v), 0, 1)
 
-    # web/src/lib/types.ts stores Field.data as Float32Array, so the browser
-    # computes on float32 risk values. Snap here -- same standing rule as
-    # poisson_fixture -- so place.py's candidate placement, demand sampling,
-    # and coverage scoring all run on the exact values the TS port will read
-    # back, instead of drifting from a float64 reference.
-    risk = risk.astype(np.float32).astype(np.float64)
+    # float32 dtype, not just float32 values -- same standing rule as
+    # poisson_fixture, and for the same NEP-50 reason documented there: the
+    # radius closure in place.py narrows to float32 only when `risk` really
+    # is float32, which is what risk.bin ships and what Field.data holds.
+    # (greedy_minimise and coverage_of both do `np.asarray(demand_w, float)`,
+    # so the weight sums stay float64 either way -- the dtype change moves
+    # the radius arithmetic only.)
+    risk = risk.astype(np.float32)
     mask = risk > 0.15
 
     w_km, h_km = 40.0, 30.0
